git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@8854 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/inet/Request.cpp
@@ -422,7 +422,7 @@ namespace inet {
         }
         if (myHeader.size()) {
             if (myRequest->onResponseHeader(myHeader) == false) {
-                return -1; // abort transfer
+                return (size_t) -1; // abort transfer
             }
         }
         return theBlockCount*theBlockSize;
